requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
numpy==2.2.1
faiss-cpu==1.9.0.post1
python-multipart==0.0.20
streamlit==1.41.1
requests==2.32.3

# Optional (recommended for local embeddings):
sentence-transformers==3.3.1
torch==2.5.1

# Optional (for OpenAI):
openai==1.59.6
python-dotenv==1.0.1
